Heap analysis: use symbolic dereferencing when accessing objects in the heap.
We transform p->n into p'obj.n which represents a concretised object.
In the same time, changes are reflected on actual heap objects
(dynamic_object$...).
This is because dynamic_object is an abstraction and so the analysis
might be
unsound when dealing with 2 different objects represented by the same
dynamic_object.
We also introduce simple may_alias analysis here.