Add new abstract domain for computing invariants for different symbolic paths
A value in the domain is a set of invariants in the compound domain (in this case
the heap interval domain). Each invariant holds for a different symbolic path in
the analysed program. We infer an invariant for each path separately, using the
possibility to restrict solving to certain paths only.
After computation of an invariant, we check whether the corresponding symbolic
path is feasible and if not, we remove the invariant.