Dynamic objects: get rid of allocation guards in templates.
Allocation guards of dynamic objects were a part of template row post guards in
order to avoid a non-deterministic to get into the template value.
We replace them by inequality saying that the given row expression (loop-back
field of a dynamic object) does not equal the same field of the same dynamic
object without suffix (representing the value of the field when the object is
not allocated).
This allows us to get rid of tracking assignments in the ssa_domain analysis.