Create fresh symbols for dynamically allocated objects and their fields.
For each object allocated at some location, create a fresh symbol (i.e. a
declaration) for it or for all of its fields in case of a structured object.
These fresh symbols can be used on rhs only if the object was allocated at the
given run - for each RHS usage, add case split. If the allocation guard holds,
use the fresh symbol, otherwise use the last definition of the object before
the allocation.