tpolyhedra: prevent using pre-value to update itself in symbolic value system.
This causes a problem for dynamically allocated object, since the value of some
x#lbN being constraint from one side (in intervals) can be used to compute the
other side of the interval. This may happen if the updated row is the last not
yet updated. In such situation, in the symbolic value system, there will be no
g#lsX => FALSE (since the last FALSE is replaced by a symbolic value), hence
x#lbN can be defined (by setting g#lsX to true) and since it is yet constraint
from one side only, it may cause the other side of the interval to be set to a
non-deterministic value.