Array domain
The domain is based on splitting the array into several segments and
then inferring a single invariant for each segment that holds for all
elements of the segment.

The segmentation is done at the beginning of the analysis based on a
newly introduced array index analysis. The analysis collects all
expressions that occur as array indices, distinguishing between read
(occurring on RHS) and write (occurring on LHS) accesses.

Written indices are used for segmentation. If all written indices (say
i_1, ..., i_n) for an array can be ordered (this is determined using the
SMT solver), the array is split into the following segments:

  {0} ... {i_1} ... {i_1+1} [...] {i_n} ... {i_n+1} ... {array size}

If there is no unique ordering over the indices, there are multiple
segmentations created (one for each written index):

  {0} ... {i_1} ... {i_1+1} ... {array size}
  [...]
  {0} ... {i_n} ... {i_n+1} ... {array size}

In addition, a segment representing the entire array is always added.
Note that the segmentation is unique for each loop and depends only on
indices written inside that loop.

For each segment S, two fresh variables are introduced:
- index variable idx#S representing all array indices within the segment
- element variable elem#S representing all array elements (i.e.
  array[idx#S]) within the segment.
Also, each segment has a lower (low#S) and upper (upp#S) bound.

The template for a segment has the (simplified) form:

  idx#S >= low#S && idx#S < upp#S && elem#S = arr[idx#S] => invariant for elem#S

where the invariant can be computed in any existing domain. At the
moment, we support the interval domain only.

After the invariants are computed, they are projected onto all read
indices (obtained from the array index analysis). There are two
approaches to the binding:
- adding equality between idx#S and a read index if there is a single
  read index for the given array or
- copying the invariant for each read index and replacing idx#S by the
  read index otherwise.
The projection must be done during invariant inference, too, to have the
context in which the invariant is solved.