Remove support for interprocedural heap analysis
The analysis is not used since it does not work. Therefore,
ssa_heap_analysis can be removed. ssa_heap_analysis was a static
analysis used to determine dynamic objects that may have been
created/altered by a function call and based on the results of this
analysis, appropriate abstractions were created.