Generic product combination abstract domain
The domain can contain multiple arbitrary abstract domains that are used
side-by-side. The invariant for each domain is computed individually in
parallel with invariants for other domains. Moreover, candidate
invariant (abstract value computed so far) from one domain is used as a
context to strengthen invariant inference in all other domains.

Heap-tpolyhedra domain is completely replaced by the new product domain.

There is an important addition to the general abstract value interface -
each abstract value needs to implement a clone method with covariant
return type. This is especially needed for the product domain so that it
can be properly cloned (clone is used by domain with symbolic paths).