Allow to use any combination of simple domains
The 2LS interface now allows to specify multiple simple domains to use.
These include --heap, --equalities, and all value domains. If mutliple
domains are specified, a product domain is created and invariants are
inferred in all domains in parallel.

Since the domains are created in template_generator_base while the
strategy solvers and abstract values are created in ssa_analyzer, it is
required that the order of simple domains/solvers/values within the
product domain/solver/value is always the same.

CLI options of all tests are ported to the new interface.