Use unique_ptr for domains and strategy solvers
Instead of using raw pointers and managing memory manually, start using
unique_ptr. If an object contains a unique_ptr to another object as a
member, it is the owner.
In particular:
- template generator owns the abstract domain
- ssa analyzer owns the abstract value
- ssa analyzer owns the strategy solver (as a local variable in
  operator())
- product domain owns the inner domains, likewise for the product value
  and for the product strategy solver
- sympath domain owns the inner domain, likewise for the sympath solver
- sympath value owns all the inner values for individual symbolic paths,
  it also owns the inner value template

Generic abstract value has an abstract method "clone" that returns a
covariant type. Since this cannot be implemented using unique_ptr, it
returns a raw pointer. The caller is therefore responsible for taking
ownership of the cloned object.