Arrays: always use symbolic size
If an array has a constant size, create a new variable, set its value to
the array size, and replace the array size by the new variable. The RHS
variant of the new variable is then used as an array size in the array
template.

This is necessary for large array sizes as the SMT solver tries to
create a model of the entire array if the size is constant.

This allows to verify programs with huge arrays. Enlarge all arrays in
regression/arrays to the size 1000000.