Move namespace ownership to local_SSA
Previously, namespace was a local variable of
summary_checker::operator() and was passed to SSA only as a reference.
However, with GOTO code instrumentation, namespace may be used once
summary_checker::operator() has finished, for example inside heap domain
when obtaining the invariant expressions. Move the ownership of
namespace to SSA so that it can be used during instrumentation.

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>