Arrays: improve invariant projection
When an array invariant is projected onto multiple read indices, it is
not sufficient to only create bindings between segment index variables
and read indices as that would force all read indices witihin the same
segment to have the same value. Instead, we have to copy the invariant
for each read index and replace the segment index variables by the read
index.