Introduce expression dependence analysis
The analysis determines which expressions (in particular variables,
array accesses, and struct field accesses) have a potential dependency
between them. We distinguish two kinds of dependencies:
- data dependency: the expressions occur inside the same
  expression/statement
- control dependency: one of the expressions occur in an if/loop
  condition and the other one is written inside that if/loop

More specifically, the analysis creates an equivalence partitioning
based on the union of the above dependency relations.

In addition, some symbols are expressions as they are known to have
non-deterministic values, and hence no reasonable dependence on other
symbols. The ignored expressions are:
- symbols that contain the "__VERIFIER_nondet"
- dynamic objects at the point of their allocation as they have
  non-deterministic content at that point

The results of the analysis can (and will) be used to make the zones
domain template smaller (by including only differences between dependent
expressions).