Arrays: use two segment element variables
We need to use different segment element variables for pre- and post-
variants of the invariant. This is needed as the elements may be
generally different, even though they access the same index. This is
because the pre-variant accesses the loop-back array, while the
post-variant accesses the loop end array).