Arrays: mark segments indices as read indices
When projecting loop invariant onto read indices, we need to project it
onto segment index variables of all segments that use the same array in
a different loop.

This is particularly necessary for situations where segments are not
updated in order (e.g. by random access). See the regression test added
in this commit for example.

Since the invariant may be quite large, which will slow down the solver,
we do this only if the zones domain is not in use. This is a safe thing
to do b/c the projection just makes the analysis more precise.