Drop usage of sympath domain
Replace symbolic paths by adding a loop guard directly to the invariant
of the heap domain. If a pointer may point to a dynamic object, add
(conjunct) a loop guard corresponding to that object to the
pointer-object equality.

That is, if the original row value was e.g.:

    p#lb50 == &dynamic_object$16 || p#lb50 == &dynamic_object$32

where dynamic_object$16 is allocated in a loop with $guard#ls30
  and dynamic_object$32 is allocated in a loop with $guard#ls50

the new row value is:

    (p#lb50 == &dynamic_object$16 && $guard#ls30) ||
    (p#lb50 == &dynamic_object$32 && $guard#ls50)

This makes sure that invariants for the dynamic object are used only if
the objects were actually allocated.