malloc: fix getting malloc size
After the rebase to new CBMC, constant propagation does not work in some
cases. This is a problem for our malloc handling as we require to see:

  malloc_size = sizeof(...)

This commit fixes the problem by recursively searching for the malloc
size expression to handle cases like:

  size = sizeof(...)
  malloc_size = size

Also adds a regression test for this case.