Fix assert to limit array size
CBMC has changed the representation of numerical values in
constant_expr (at least for array sizes) - instead of base 2, it now
uses base 16.

We have to fix this in our assertion that limits the size of arrays in
competition mode due to solver unsoundness that appears for arrays of
size >=50000.