Dynamic objects: collect loop guards
For each dynamic object, collect the loop-select guard of the loop that
it is allocated in (or a conjunction of multiple guards in case the
object is allocated in a nested loop).

The guards will be used to replace symbolic paths as they must hold if
the object has been allocated (and therefore its invariant is valid).