Domains: do not include top rows into invariant
When doing projection of an invariant onto variables, ignore rows whose
abstract value is equal to the top of the given abstract domain. This
will make the invariants considerably smaller and speed up solving when
an invariant must be often put to the solver (as is the case for the
array domain which uses the current invariants during each strategy
solving iteration).