Remove array size limitation in competition mode
Commit ("461e970" Temporary hack: disable arrays larger than 50000.)
introduced a hack to disable arrays larger than 50000 items in
competition mode. The reason for this was that the SMT solver was not
able to create sound models of such large arrays.

Since we now make the array sizes symbolic, this problem should be
mitigated.