Fix GOTO undinwing of dynamic objects
Commit 89e30e27 ("Get rid of the "#malloc_result" property") replaced
usage of the #malloc_result property by querying the dynamic objects
database. This is not sufficient for the GOTO unwinder since it used to
use that property to locate the cloned malloc instruction (already
containing the generated dynamic objects). As the GOTO unwinder does not
duplicate entries in the dynamic objects database, the malloc location
is now not identified properly.

This fixes the above problem by introducing a #malloc property, local to
the unwinder, which we use to mark the malloc result instruction prior
to running the CBMC unwinder. Since properties are cloned during
unwinding, we can use it to locate the new malloc instruction and
properly generate new dynamic objects.

Add test from SV-COMP which contains a linked list of a fixed length, so
can be verified with sufficient unwinding without computing invariants.