Support arrays with standard invariants
When --std-invariants is set, invariants are computed for the loop head
(i.e. contain the pre-loop value in addition to the loop-back values).
This is not convenient for arrays since they are typically uninitialized
and only get initialized inside a loop. Including the pre-loop array
value would add the nondet value to the invariant, making the entire
invariant (especially the one for the initializer loop) nondet.

This change prohibits usage of standard invariants for the array domain.
Most of the changes are done in the template generator which:
- does not generate aux_expr for array var specs,
- always passes post_renaming_map (instead of aux_renaming_map) to the
  array domain and its inner domains.